requirements.txt
fastapi
uvicorn
pytesseract
pillow
httpx
python-multipart
pytest